manifest: Validate required data is present in manifest files (#5840)
If a manifest file does not contain a config entry (or is corrupt), Ollama will crash
will a nil pointer dereference when reading it, such as on pull, startup or listing
models. Instead, we should check that this is present when deserializing
data.

In addition, SchemaVersion is not currently verified, which can cause
problems in the future if an incompatible change needs to be made.

Other types of validation errors are already handled, such as malformed
JSON or other missing fields.